# tests/GalaxyData.Web.Tests/Hosting/Snapshots/HostTests.TheOpenApiDocumentDescribesTheApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GalaxyData",
    "description": "Browse, query and change data across databases and folders of workbooks. Errors are problem details (RFC 9457) with a code.",
    "version": "v1"
  },
  "paths": {
...
    "/api/changes": {
      "get": {
        "tags": [
          "Changes"
        ],
        "summary": "The user's pending changes, in the order they were first made",
        "operationId": "GetChanges",
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Changes"
        ],
        "summary": "Drops the pending changes, or those of a source or an entity",
        "operationId": "ClearChanges",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/ops": {
      "post": {
        "tags": [
          "Changes"
        ],
        "summary": "Applies operations to the pending changes: all of them, or none",
        "operationId": "ApplyChangeOps",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOpsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeSetDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/preview": {
      "post": {
        "tags": [
          "Changes"
        ],
        "summary": "The statements committing the changes would run on each connection, and why changes can't be made; a plan to commit when none",
        "operationId": "PreviewChanges",
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePreviewDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/changes/commit": {
      "post": {
        "tags": [
          "Changes"
        ],
        "summary": "Commits a preview's plan, with scripts as edited; what came of it on each connection",
        "operationId": "CommitChanges",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitChangesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChangeIssueDto": {
        "required": [
          "change",
          "column",
          "message"
        ],
        "type": "object",
        "properties": {
          "change": {
            "type": "integer",
            "format": "int64"
          },
          "column": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ChangeOpDto": {
        "required": [
          "op"
        ],
        "type": "object",
        "properties": {
          "op": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ChangeOpKind"
              }
            ]
          },
          "entity": {
            "type": [
              "null",
              "string"
            ]
          },
          "key": {
            "type": [
              "null",
              "array"
            ]
          },
          "tempId": {
            "type": [
              "null",
              "string"
            ]
          },
          "change": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "values": {
            "type": [
              "null",
              "object"
            ]
          },
          "original": {
            "type": [
              "null",
              "object"
            ]
          },
          "display": {
            "type": [
              "null",
              "object"
            ]
          },
          "columns": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ChangeOpKind": {
        "enum": [
          "set",
          "insert",
          "delete",
          "revert"
        ]
      },
      "ChangeOpsRequest": {
        "required": [
          "ops"
        ],
        "type": "object",
        "properties": {
          "ops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChangeOpDto"
            }
          },
          "version": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
...
      "ChangePreviewDto": {
        "required": [
          "planId",
          "version",
          "catalogVersion",
          "expiresAt",
          "multiConnection",
          "scripts",
          "issues"
        ],
        "type": "object",
        "properties": {
          "planId": {
            "type": [
              "null",
              "string"
            ]
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "catalogVersion": {
            "type": "string"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "multiConnection": {
            "type": "boolean"
          },
          "scripts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewScriptDto"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChangeIssueDto"
            }
          }
        }
      },
      "ChangeSetDto": {
        "required": [
          "version",
          "updatedAt",
          "changes"
        ],
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingChangeDto"
            }
          }
        }
      },
...
      "CommitChangesRequest": {
        "required": [
          "planId",
          "version"
        ],
        "type": "object",
        "properties": {
          "planId": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "scripts": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/EditedScriptDto"
            }
          },
          "allowAnyStatement": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CommitFailureDto": {
        "required": [
          "kind",
          "source",
          "message",
          "change",
          "statement"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DmlFailureKind"
          },
          "source": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "change": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "statement": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CommitResultDto": {
        "required": [
          "auditId",
          "outcome",
          "scripts",
          "failure",
          "inserted",
          "changes",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "auditId": {
            "type": "integer",
            "format": "int64"
          },
          "outcome": {
            "$ref": "#/components/schemas/DmlOutcome"
          },
          "scripts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitScriptDto"
            }
          },
          "failure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CommitFailureDto"
              }
            ]
          },
          "inserted": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InsertedRowDto"
            }
          },
          "changes": {
            "$ref": "#/components/schemas/ChangeSetDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CommitScriptDto": {
        "required": [
          "source",
          "edited",
          "status",
          "error",
          "statements"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "edited": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DmlScriptStatus"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "statements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitStatementDto"
            }
          }
        }
      },
...
      "CommitStatementDto": {
        "required": [
          "change",
          "description",
          "rowsChanged"
        ],
        "type": "object",
        "properties": {
          "change": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "rowsChanged": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "DmlFailureKind": {
        "enum": [
          "connection",
          "statement",
          "conflict",
          "commit",
          "timeout"
        ]
      },
      "DmlOutcome": {
        "enum": [
          "committed",
          "rolledBack",
          "partiallyCommitted"
        ]
      },
      "DmlScriptStatus": {
        "enum": [
          "committed",
          "rolledBack",
          "commitFailed"
        ]
      },
      "DmlStatementKind": {
        "enum": [
          "insert",
          "update",
          "delete",
          "merge",
          "other"
        ]
      },
      "EditedScriptDto": {
        "required": [
          "source",
          "text"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "InsertedRowDto": {
        "required": [
          "tempId",
          "entity",
          "key",
          "rowId",
          "values"
        ],
        "type": "object",
        "properties": {
          "tempId": {
            "type": "string"
          },
          "entity": {
            "type": "string"
          },
          "key": {
            "type": [
              "null",
              "array"
            ]
          },
          "rowId": {
            "type": [
              "null",
              "string"
            ]
          },
          "values": {
            "type": "object"
          }
        }
      },
...
      "PendingChangeDto": {
        "required": [
          "id",
          "kind",
          "entity",
          "source",
          "key",
          "rowId",
          "tempId",
          "values",
          "original",
          "display",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kind": {
            "$ref": "#/components/schemas/PendingChangeKind"
          },
          "entity": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "key": {
            "type": [
              "null",
              "array"
            ]
          },
          "rowId": {
            "type": [
              "null",
              "string"
            ]
          },
          "tempId": {
            "type": [
              "null",
              "string"
            ]
          },
          "values": {
            "type": "object"
          },
          "original": {
            "type": "object"
          },
          "display": {
            "type": "object"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PendingChangeKind": {
        "enum": [
          "insert",
          "update",
          "delete"
        ]
      },
...
      "PreviewScriptDto": {
        "required": [
          "source",
          "kind",
          "dialect",
          "text",
          "statements"
        ],
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "dialect": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "statements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreviewStatementDto"
            }
          }
        }
      },
      "PreviewStatementDto": {
        "required": [
          "change",
          "kind",
          "description",
          "text"
        ],
        "type": "object",
        "properties": {
          "change": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "kind": {
            "$ref": "#/components/schemas/DmlStatementKind"
          },
          "description": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string",
            "description": "What the problem is, for clients to tell problems apart by"
          },
          "traceId": {
            "type": "string",
            "description": "The request's trace, to find it in the logs"
          }
        }
      },
...
    }
  },
...
}